refactor: Refactor files in the backend folder.
